# docs/swagger/swagger.yaml
basePath: /api/v2
definitions:
...
  common.JSONResult:
    properties:
      data:
        description: 数据
      extra:
        description: 数据
      msg:
        description: 消息
        type: string
      success:
        description: 是否成功
        type: boolean
    type: object
...
  entity.CookedUserForAdmin:
    properties:
      badge_color:
        type: string
      badge_name:
        type: string
      comment_count:
        type: integer
      email:
        type: string
      id:
        type: integer
      is_admin:
        type: boolean
      is_in_conf:
        type: boolean
      last_ip:
        type: string
      last_ua:
        type: string
      link:
        type: string
      name:
        type: string
      receive_email:
        type: boolean
      site_names:
        items:
          type: string
        type: array
      site_names_raw:
        type: string
    type: object
...
  handler.ParamsAdminImport:
    properties:
      payload:
        description: The transfer importer payload
        type: string
    type: object
...
  handler.ParamsAdminUserEdit:
    properties:
      badge_color:
        description: The user badge color (hex format)
        type: string
      badge_name:
        description: The user badge name
        type: string
      email:
        description: The user email
        type: string
      is_admin:
        description: The user is an admin
        type: boolean
      link:
        description: The user link
        type: string
      name:
        description: The user name
        type: string
      password:
        description: The user password
        type: string
      receive_email:
        description: The user receive email
        type: boolean
      site_names:
        description: The site names associated with the user
        type: string
    required:
    - email
    - is_admin
    - name
    - receive_email
    type: object
  handler.ParamsCaptchaCheck:
    properties:
      value:
        description: The captcha value to check
        type: string
    required:
    - value
    type: object
...
  handler.ResponseAdminUserEdit:
    properties:
      user:
        $ref: '#/definitions/entity.CookedUserForAdmin'
    type: object
  handler.ResponseAdminUserGet:
    properties:
      total:
        type: integer
      users:
        items:
          $ref: '#/definitions/entity.CookedUserForAdmin'
        type: array
    type: object
...
info:
  contact:
    email: artalkjs@gmail.com
    name: API Support
    url: https://artalk.js.org
  description: Artalk is a modern comment system based on Golang.
  license:
    name: MIT
    url: https://github.com/ArtalkJS/Artalk/blob/master/LICENSE
  title: Artalk API
  version: "2.0"
paths:
  /artransfer/export:
    get:
      description: Export data from Artalk
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.JSONResult'
            - properties:
                data:
                  properties:
                    data:
                      type: string
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Export Artrans
      tags:
      - Transfer
  /artransfer/import:
    post:
      consumes:
      - application/json
      description: Import data to Artalk
      parameters:
      - description: The data to import
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/handler.ParamsAdminImport'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.JSONResult'
      security:
      - ApiKeyAuth: []
      summary: Import Artrans
      tags:
      - Transfer
  /artransfer/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload a file to prepare to import
      parameters:
      - description: Upload file in preparation for import task
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.JSONResult'
            - properties:
                data:
                  properties:
                    filename:
                      type: string
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Upload Artrans
      tags:
      - Transfer
...
  /captcha/verify:
    post:
      description: Verify user enters correct captcha code
      parameters:
      - description: The data to check
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/handler.ParamsCaptchaCheck'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.JSONResult'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/common.JSONResult'
            - properties:
                data:
                  properties:
                    img_data:
                      type: string
                  type: object
              type: object
      summary: Verify Captcha
      tags:
      - Captcha
...
  /users/{id}:
    delete:
      description: Delete a specific user
      parameters:
      - description: The user ID you want to delete
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.JSONResult'
      security:
      - ApiKeyAuth: []
      summary: Delete User
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Edit a specific user
      parameters:
      - description: The user ID you want to edit
        in: path
        name: id
        required: true
        type: integer
      - description: The user data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handler.ParamsAdminUserEdit'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/handler.ResponseAdminUserEdit'
              type: object
      security:
      - ApiKeyAuth: []
      summary: Edit User
      tags:
      - User
  /users/{type}:
    get:
      consumes:
      - application/json
      description: Get a list of users by some conditions
      parameters:
      - description: The type of users
        enum:
        - all
        - admin
        - in_conf
        in: path
        name: type
        type: string
      - description: The limit for pagination
        in: query
        name: limit
        type: integer
      - description: The offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/handler.ResponseAdminUserGet'
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get User List
      tags:
      - User
...
securityDefinitions:
  ApiKeyAuth:
    description: '"Type ''Bearer TOKEN'' to correctly set the API Key"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"